iso9660: initial commit
* add U32 and U16 macros to the byteorder.h for an easy way of defining
  signatures and other values out of single bytes
* change buffer.data's type to void*
* try to mount devices with iso9660 and ext2 in init
* add shift and shift_as macros to compiler.h
* make ino_t an unsigned long

+++ include/kernel/types.h
@@ -4,7 +4,7 @@ typedef unsigned short pid_t;
 typedef char pgid_t;
 typedef unsigned int size_t;
 typedef unsigned int off_t;
-typedef unsigned short ino_t;
+typedef unsigned long ino_t;
 typedef int daddr_t;
 typedef unsigned short kdev_t;
 typedef unsigned short mode_t;
